add get rgb components for image at point

+++ GPUImageDemo/GPUImageDemo/Class/Makeup/FaceLandmark/ncnn.framework/Versions/A/Headers/mat.h
@@ -289,7 +289,7 @@ inline Mat& Mat::operator=(const Mat& m)
 
 inline void Mat::fill(float _v)
 {
-    int size = total();
+    int size = (int)total();
     float* ptr = (float*)data;
 
 #if __ARM_NEON
@@ -343,7 +343,7 @@ inline void Mat::fill(float _v)
 
 inline void Mat::fill(int _v)
 {
-    int size = total();
+    int size = (int)total();
     int* ptr = (int*)data;
 
 #if __ARM_NEON
@@ -398,7 +398,7 @@ inline void Mat::fill(int _v)
 template <typename T>
 inline void Mat::fill(T _v)
 {
-    int size = total();
+    int size = (int)total();
     T* ptr = (T*)data;
     for (int i=0; i<size; i++)
     {
